fix: 非同期処理（multer）によるテナントコンテキスト消失の修正 (#46-9)
