+++ backend/src/server.ts
@@ -1,21 +1,16 @@
 import express, { Request, Response, NextFunction } from 'express';
-import multer from 'multer';
 import cors from 'cors';
 import path from 'path';
 import fs from 'fs';
 import dotenv from 'dotenv';
-import sharp from 'sharp'; 
 
 // --- [Issue #43] BullMQ & Redis ---
-import { receiptQueue } from './queues/receiptQueue';
 import './workers/receiptWorker'; 
 
 // --- ユーティリティ & ミドルウェア ---
 import logger from './utils/logger';
-import { prisma } from './utils/prismaClient'; 
 import { authMiddleware } from './middleware/authMiddleware'; 
 import { tenantMiddleware } from './middleware/tenantMiddleware';
-import { getFamilyGroupId, getMemberId } from './utils/context';
 
 // --- ルーター ---
 import authRoutes from './routes/authRoutes'; 
@@ -26,8 +21,6 @@ import categoryRoutes from './routes/categoryRoutes';
 // --- エラーハンドリング ---
 import { AppError } from './utils/appError';
 import { errorHandler } from './middleware/errorHandler';
-import { validate } from './middleware/validate';
-import { uploadReceiptSchema } from './schemas/receiptSchema';
 
 dotenv.config();
 
@@ -55,68 +48,29 @@ app.use(cors({
 }));
 
 app.use(express.json());
-app.use('/uploads', express.static(path.join(process.cwd(), 'uploads')));
 
-// --- 2. [Issue #52] ルート登録 ---
-app.use('/api/auth', authRoutes);
-app.use('/api', authMiddleware, tenantMiddleware);
-app.use('/api/categories', categoryRoutes);
-app.use('/api/product-master', productMasterRoutes);
-app.use('/api', receiptRoutes);
-
-// --- 3. ストレージ設定 ---
-const uploadDir = 'uploads';
+// 静的ファイルの提供
+const uploadDir = path.join(process.cwd(), 'uploads');
 if (!fs.existsSync(uploadDir)) {
-  fs.mkdirSync(uploadDir);
+  fs.mkdirSync(uploadDir, { recursive: true });
 }
-const storage = multer.memoryStorage();
-const upload = multer({ storage: storage });
+app.use('/uploads', express.static(uploadDir));
 
-app.post(
-  '/api/receipts/upload', 
-  upload.single('image'), 
-  validate(uploadReceiptSchema),
-  async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const file = req.file as Express.Multer.File; 
-      if (!file) throw new AppError('画像がアップロードされていません。', 400);
+// --- 2. ルート登録 ---
 
-      const familyGroupId = getFamilyGroupId();
-      const memberId = getMemberId(); 
+// 認証なしルート
+app.use('/api/auth', authRoutes);
 
-      if (!memberId || !familyGroupId) {
-        throw new AppError('認証コンテキストが不正です。', 401);
-      }
+/**
+ * [重要] 認証・テナント必須ルート
+ * ここで一括適用せず、各ルーターに任せるか、あるいはパスを指定して適用します。
+ * 今回はコンテキスト消失を防ぐため、receiptRoutes側で個別に制御できるようにします。
+ */
+app.use('/api', receiptRoutes);
+app.use('/api/categories', authMiddleware, tenantMiddleware, categoryRoutes);
+app.use('/api/product-master', authMiddleware, tenantMiddleware, productMasterRoutes);
 
-      const timestamp = Date.now();
-      const baseFileName = `receipt-${timestamp}-${Math.round(Math.random() * 1e9)}`;
-      const imagePath = path.join(uploadDir, `${baseFileName}.webp`);
-
-      await sharp(file.buffer)
-        .rotate() 
-        .resize(1000, 1000, { fit: 'inside', withoutEnlargement: true }) 
-        .webp({ quality: 75, effort: 6 }) 
-        .toFile(imagePath);
-
-      const job = await receiptQueue.add('analyze-receipt', {
-        memberId: memberId,
-        familyGroupId: familyGroupId,
-        imagePath: imagePath,
-      });
-
-      logger.info(`[Queue] 解析ジョブ登録: ID ${job.id} (世帯: ${familyGroupId}, 会員: ${memberId})`);
-
-      res.status(202).json({
-        success: true,
-        data: { jobId: job.id, status: 'queued' }
-      });
-      
-    } catch (error) {
-      next(error);
-    }
-  }
-);
-
+// 404 & Error Handler
 app.use((req, res, next) => next(new AppError(`Not Found - ${req.originalUrl}`, 404)));
 app.use(errorHandler);
